refactor(web): move `Pricing` component from home to shared
